Protect create route
